Renames for best practice and cleaner interface understanding
